fix: relocate general graphql
