+++ pages/statemine/u/_id.vue
@@ -141,10 +141,10 @@ import shortAddress from '@/utils/shortAddress'
 import collectionList from '@/queries/unique/collectionListByAccount.graphql'
 import PrefixMixin from '@/utils/mixins/prefixMixin'
 
-import firstNftByIssuer from '@/queries/rmrk/subsquid/general/firstNftByIssuer.graphql'
-import nftListByIssuer from '@/queries/rmrk/subsquid/general/nftListByIssuer.graphql'
-import nftListCollected from '@/queries/rmrk/subsquid/general/nftListCollected.graphql'
-import nftListSold from '@/queries/rmrk/subsquid/general/nftListSold.graphql'
+import firstNftByIssuer from '@/queries/subsquid/general/firstNftByIssuer.graphql'
+import nftListByIssuer from '@/queries/subsquid/general/nftListByIssuer.graphql'
+import nftListCollected from '@/queries/subsquid/general/nftListCollected.graphql'
+import nftListSold from '@/queries/subsquid/general/nftListSold.graphql'
 
 const components = {
   GalleryCardList: () =>
