+++ queries/subsquid/general/firstNftByIssuer.graphql
